+++ datanode/sqlsubscribers/positions_test.go
@@ -459,7 +459,7 @@ func TestPositionSpecSuite(t *testing.T) {
 			ps := tc.pos
 			sub, store := getSubscriberAndStore(t)
 			sub.Push(context.Background(), ps)
-			pp, err := store.GetByMarket(ctx, entities.MarketID(market))
+			pp, err := store.GetByMarket(ctx, market)
 			assert.NoError(t, err)
 			assert.NotZero(t, len(pp))
 			// average entry price should be 1k
@@ -483,11 +483,11 @@ func getSubscriberAndStore(t *testing.T) (*sqlsubscribers.Position, sqlsubscribe
 		return nil
 	}
 
-	getByMarket := func(_ context.Context, _ entities.MarketID) ([]entities.Position, error) {
+	getByMarket := func(_ context.Context, _ string) ([]entities.Position, error) {
 		return []entities.Position{lastPos}, nil
 	}
 
-	getByMarketAndParty := func(_ context.Context, _ entities.MarketID, _ entities.PartyID) (entities.Position, error) {
+	getByMarketAndParty := func(_ context.Context, _ string, _ string) (entities.Position, error) {
 		return lastPos, nil
 	}